+++ piezas.py
@@ -13,12 +13,13 @@ class Casilla:
 class Espacio(Casilla):
     pass
 
-
+# Ya no sirven num ni color_casilla.
 class Pieza(Casilla):
-    def __init__(self, var, color, posicion, s, color_casilla, nom, vive = True):
+    def __init__(self, var, color, posicion, s, nom, vive = True):
         super().__init__(var, color, s)
         self.__posicion__ = posicion
-        self.__color_casilla__ = color_casilla
+        # self.__color_casilla__ = color_casilla
+        # self.__num__ = num
         self.__vive__ = vive
         self.__nom__ = nom # Nombre de la pieza.
 
@@ -27,14 +28,14 @@ class Pieza(Casilla):
         return self.__var__
 
     # Actualiza la posición de la pieza.
-    def mover(self, nueva_posicion, nuevo_color_casilla):
+    def mover(self, nueva_posicion):
         self.__posicion__ = nueva_posicion
-        self.__color_casilla__ = nuevo_color_casilla
+        # self.__color_casilla__ = nuevo_color_casilla
 
 
 class Peon(Pieza):
-    def __init__(self, var, color, posicion, s, color_casilla, nom, primera_posicion = True):
-        super().__init__(var, color, posicion, s, color_casilla, nom)
+    def __init__(self, var, color, posicion, s, nom, primera_posicion = True):
+        super().__init__(var, color, posicion, s, nom)
         self.__primera_posicion__ = primera_posicion # Indica si es la primera vez que se mueve la pieza.
      
     def movimientos_posibles(self):
@@ -89,9 +90,9 @@ class Peon(Pieza):
         return posiciones_posibles
     
     # El método esta repetido porque acá especificamente cambia el atributo 'primera_posicion'.
-    def mover(self, nueva_posicion, nuevo_color_casilla):
+    def mover(self, nueva_posicion):
         self.__posicion__ = nueva_posicion
-        self.__color_casilla__ = nuevo_color_casilla
+        # self.__color_casilla__ = nuevo_color_casilla
         self.__primera_posicion__ = False